move descovery_message into service_discovery
